+++ 3DAuth/Util.cs
@@ -8,7 +8,7 @@ namespace ThreeDAuth
 
     class Util
     {
-        public static double euclideanDistance(Point2f p1, Point2f p2)
+        public static double euclideanDistance(Point2d p1, Point2d p2)
         {
             return Math.Sqrt((p1.X - p2.X) * (p1.X - p2.X) +
                                 (p1.Y - p2.Y) * (p1.Y - p2.Y));
